Moved compaction DAG pruing configs back to OzoneConfigKeys
