+++ hadoop-ozone/ozone-manager/src/main/java/org/apache/hadoop/ozone/om/OmSnapshotManager.java
@@ -209,75 +209,7 @@ public final class OmSnapshotManager implements AutoCloseable {
         OZONE_OM_SNAPSHOT_CACHE_MAX_SIZE,
         OZONE_OM_SNAPSHOT_CACHE_MAX_SIZE_DEFAULT);
 
-    CacheLoader<String, OmSnapshot> loader;
-    loader = new CacheLoader<String, OmSnapshot>() {
-      @Override
-
-      // load the snapshot into the cache if not already there
-      @Nonnull
-      public OmSnapshot load(@Nonnull String snapshotTableKey)
-          throws IOException {
-        SnapshotInfo snapshotInfo;
-        // see if the snapshot exists
-        snapshotInfo = getSnapshotInfo(snapshotTableKey);
-
-        // Block snapshot from loading when it is no longer active e.g. DELETED,
-        // unless this is called from SnapshotDeletingService.
-        // TODO: [SNAPSHOT] However, snapshotCache.get() from other requests
-        //  (not from SDS) would be able to piggyback off of this because
-        //  snapshot still in cache won't trigger loader again.
-        //  This needs proper addressal in e.g. HDDS-7935
-        //  by introducing another cache just for SDS.
-        //  While the snapshotCache would host ACTIVE snapshots only.
-        if (!snapshotInfo.getSnapshotStatus().equals(
-                SnapshotStatus.SNAPSHOT_ACTIVE)) {
-          if (isCalledFromSnapshotDeletingService()) {
-            LOG.debug("Permitting {} to load snapshot {} in status: {}",
-                SnapshotDeletingService.class.getSimpleName(),
-                snapshotInfo.getTableKey(),
-                snapshotInfo.getSnapshotStatus());
-          } else {
-            throw new OMException("Unable to load snapshot. " +
-                "Snapshot with table key '" + snapshotTableKey +
-                "' is no longer active", FILE_NOT_FOUND);
-          }
-        }
-
-        CacheValue<SnapshotInfo> cacheValue =
-            ozoneManager.getMetadataManager().getSnapshotInfoTable()
-                .getCacheValue(new CacheKey<>(snapshotTableKey));
-        boolean isSnapshotInCache = cacheValue != null && Optional.ofNullable(
-            cacheValue.getCacheValue()).isPresent();
-
-        // read in the snapshot
-        OzoneConfiguration conf = ozoneManager.getConfiguration();
-        OMMetadataManager snapshotMetadataManager;
-
-        // Create the snapshot metadata manager by finding the corresponding
-        // RocksDB instance, creating an OmMetadataManagerImpl instance based on
-        // that
-        try {
-          snapshotMetadataManager = new OmMetadataManagerImpl(conf,
-                  snapshotInfo.getCheckpointDirName(), isSnapshotInCache);
-        } catch (IOException e) {
-          LOG.error("Failed to retrieve snapshot: {}, {}", snapshotTableKey, e);
-          throw e;
-        }
-
-        // create the other manager instances based on snapshot metadataManager
-        PrefixManagerImpl pm = new PrefixManagerImpl(snapshotMetadataManager,
-            false);
-        KeyManagerImpl km = new KeyManagerImpl(null,
-            ozoneManager.getScmClient(), snapshotMetadataManager, conf,
-            ozoneManager.getBlockTokenSecretManager(),
-            ozoneManager.getKmsProvider(), ozoneManager.getPerfMetrics());
-
-        return new OmSnapshot(km, pm, ozoneManager,
-            snapshotInfo.getVolumeName(),
-            snapshotInfo.getBucketName(),
-            snapshotInfo.getName());
-      }
-    };
+    CacheLoader<String, OmSnapshot> loader = createLoadingCache();
 
     RemovalListener<String, OmSnapshot> removalListener
         = notification -> {
@@ -337,6 +269,77 @@ public final class OmSnapshotManager implements AutoCloseable {
         new SnapshotDiffJob.SnapshotDiffJobCodec());
 
     return registry;
+  }
+
+  private CacheLoader<String, OmSnapshot> createLoadingCache() {
+    return new CacheLoader<String, OmSnapshot>() {
+      @Override
+
+      // load the snapshot into the cache if not already there
+      @Nonnull
+      public OmSnapshot load(@Nonnull String snapshotTableKey)
+          throws IOException {
+        SnapshotInfo snapshotInfo;
+        // see if the snapshot exists
+        snapshotInfo = getSnapshotInfo(snapshotTableKey);
+
+        // Block snapshot from loading when it is no longer active e.g. DELETED,
+        // unless this is called from SnapshotDeletingService.
+        // TODO: [SNAPSHOT] However, snapshotCache.get() from other requests
+        //  (not from SDS) would be able to piggyback off of this because
+        //  snapshot still in cache won't trigger loader again.
+        //  This needs proper addressal in e.g. HDDS-7935
+        //  by introducing another cache just for SDS.
+        //  While the snapshotCache would host ACTIVE snapshots only.
+        if (!snapshotInfo.getSnapshotStatus().equals(
+            SnapshotStatus.SNAPSHOT_ACTIVE)) {
+          if (isCalledFromSnapshotDeletingService()) {
+            LOG.debug("Permitting {} to load snapshot {} in status: {}",
+                SnapshotDeletingService.class.getSimpleName(),
+                snapshotInfo.getTableKey(),
+                snapshotInfo.getSnapshotStatus());
+          } else {
+            throw new OMException("Unable to load snapshot. " +
+                "Snapshot with table key '" + snapshotTableKey +
+                "' is no longer active", FILE_NOT_FOUND);
+          }
+        }
+
+        CacheValue<SnapshotInfo> cacheValue =
+            ozoneManager.getMetadataManager().getSnapshotInfoTable()
+                .getCacheValue(new CacheKey<>(snapshotTableKey));
+        boolean isSnapshotInCache = cacheValue != null && Optional.ofNullable(
+            cacheValue.getCacheValue()).isPresent();
+
+        // read in the snapshot
+        OzoneConfiguration conf = ozoneManager.getConfiguration();
+        OMMetadataManager snapshotMetadataManager;
+
+        // Create the snapshot metadata manager by finding the corresponding
+        // RocksDB instance, creating an OmMetadataManagerImpl instance based on
+        // that
+        try {
+          snapshotMetadataManager = new OmMetadataManagerImpl(conf,
+              snapshotInfo.getCheckpointDirName(), isSnapshotInCache);
+        } catch (IOException e) {
+          LOG.error("Failed to retrieve snapshot: {}, {}", snapshotTableKey, e);
+          throw e;
+        }
+
+        // create the other manager instances based on snapshot metadataManager
+        PrefixManagerImpl pm = new PrefixManagerImpl(snapshotMetadataManager,
+            false);
+        KeyManagerImpl km = new KeyManagerImpl(null,
+            ozoneManager.getScmClient(), snapshotMetadataManager, conf,
+            ozoneManager.getBlockTokenSecretManager(),
+            ozoneManager.getKmsProvider(), ozoneManager.getPerfMetrics());
+
+        return new OmSnapshot(km, pm, ozoneManager,
+            snapshotInfo.getVolumeName(),
+            snapshotInfo.getBucketName(),
+            snapshotInfo.getName());
+      }
+    };
   }
 
   /**
